+++ app/src/main/java/org/simple/clinic/home/overdue/search/OverdueSearchEvent.kt
@@ -3,6 +3,7 @@ package org.simple.clinic.home.overdue.search
 import androidx.paging.PagingData
 import org.simple.clinic.home.overdue.OverdueAppointment
 import org.simple.clinic.home.overdue.search.OverdueSearchQueryValidator.Result
+import java.util.UUID
 
 sealed class OverdueSearchEvent
 
@@ -13,3 +14,7 @@ data class OverdueSearchQueryChanged(val searchQuery: String) : OverdueSearchEve
 data class OverdueSearchQueryValidated(val result: Result) : OverdueSearchEvent()
 
 data class OverdueSearchResultsLoaded(val overdueAppointments: PagingData<OverdueAppointment>) : OverdueSearchEvent()
+
+data class OverduePatientClicked(val patientUuid: UUID) : OverdueSearchEvent()
+
+data class CallPatientClicked(val patientUuid: UUID) : OverdueSearchEvent()